requirements.txt
Flask==3.1.3
openai-whisper
sentence-transformers
torch
numpy
Werkzeug==3.1.6
Jinja2==3.1.6
blinker==1.9.0
click==8.3.1
colorama==0.4.6
itsdangerous==2.2.0
MarkupSafe==3.0.3